Extracts zipkin-tests jar before we publish to o.a.z.zipkin2 namespace
This fixes an anti-pattern before we commit to it in the new namespace.
