+++ pom.xml
@@ -33,6 +33,7 @@
 
   <modules>
     <module>zipkin</module>
+    <module>zipkin-tests</module>
     <module>zipkin-ui</module>
     <module>zipkin-lens</module>
     <module>zipkin-junit</module>
