+++ zipkin-server/pom.xml
@@ -295,34 +295,9 @@
     </dependency>
 
     <dependency>
-      <groupId>${project.groupId}.zipkin2</groupId>
-      <artifactId>zipkin</artifactId>
+      <groupId>org.apache.zipkin.zipkin2</groupId>
+      <artifactId>zipkin-tests</artifactId>
       <version>${project.version}</version>
-      <type>test-jar</type>
-      <scope>test</scope>
-    </dependency>
-    <!-- While gson is handled by shade, there's an edge case when ./mvnw test
-         is used in this multi-module build (ex via travis). This adds the dep
-         so that tests execute properly. Explanation below.
-
-         We shade the internal classes used in the test jar, and use optional
-         scope to prevent other modules from accidentally pulling in gson. This
-         works except that the test-jar when used here via ./mvnw test. Unlike
-         3rd party usage, which would resolve to zipkin2:zipkin's shaded dep,
-         ./mvnw test here will result in a project reference, and ignore the
-         gson dependency because it is set as optional. Note this doesn't occur
-         when ./mvnw install is used, only ./mvnw test. This may be a bug in
-         the way dependencies resolve, where some paths consider the shade
-         plugin and others don't.
-
-         This problem is actually more that we didn't initially create a -tests
-         jar, like we do in Brave. Before we release into apache, we should
-         remove the test-jar and make a zipkin-tests artifact instead, which
-         eliminates this edge case and follows recommended practice.
-      -->
-    <dependency>
-      <groupId>com.google.code.gson</groupId>
-      <artifactId>gson</artifactId>
       <scope>test</scope>
     </dependency>
