+++ zipkin-tests/src/main/java/zipkin2/TestObjects.java
@@ -22,11 +22,9 @@ import java.util.List;
 import java.util.Random;
 import java.util.TimeZone;
 import java.util.concurrent.TimeUnit;
-import zipkin2.storage.QueryRequest;
 
 import static java.util.Arrays.asList;
 
-// re-declared in zipkin-tests to avoid having to move most tests out of this module
 public final class TestObjects {
   public static final Charset UTF_8 = Charset.forName("UTF-8");
   /** Notably, the cassandra implementation has day granularity */
@@ -104,6 +102,7 @@ public final class TestObjects {
   // storage query units are milliseconds, while trace data is microseconds
   public static final long TRACE_DURATION = TRACE.get(0).durationAsLong() / 1000;
   public static final long TRACE_STARTTS = TRACE.get(0).timestampAsLong() / 1000;
+  public static final long TRACE_ENDTS = TRACE_STARTTS + TRACE_DURATION;
 
   static final Span.Builder spanBuilder = spanBuilder();
 
@@ -127,9 +126,5 @@ public final class TestObjects {
 
   public static Span span(long traceId) {
     return spanBuilder.traceId(Long.toHexString(traceId)).id(traceId).build();
-  }
-
-  public static QueryRequest.Builder requestBuilder() {
-    return QueryRequest.newBuilder().endTs(TODAY + DAY).lookback(DAY * 2).limit(100);
   }
 }
