+++ zipkin-tests/src/main/java/zipkin2/storage/ITAutocompleteTags.java
